Add raw page allocator
