+++ kernel/Makefile
@@ -1,7 +1,7 @@
 export CC CFLAGS CPPFLAGS LDFLAGS LIBS
 OBJECTS=./*.o ./x86/*.o ./x86/asm/*.o ./lib/*.o ./drivers/*.o
 #TODO: Make following list two seperate lists
-all: kernel.o panic.o slaballoc.o mm.o timer.o x86 lib drivers
+all: kernel.o panic.o slaballoc.o mm.o page_alloc.o timer.o x86 lib drivers
 	@echo "Building kernel image"
 	$(CC) -T ./linker.ld -o ../sysroot/boot/kernel -ffreestanding -O2 -nostdlib $(OBJECTS) -lgcc
