+++ kernel/x86/asm/boot.S
@@ -30,7 +30,7 @@ kernelpagedir:
 	.long 0x00000083
 	.long 0x00400083
 	.long 0x00800083
-	.skip 1024-KERNEL_PAGE_INDEX-3
+	.skip (1024-KERNEL_PAGE_INDEX-3)*4
 //TODO:Once userland init starts, system should use that process's page directory. The page directory of every process should have the kernel mapped to 0xC000000 and above.
 
 .section .text
